+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+<a name="0.1.2"></a>
+## [0.1.2](https://github.com/wswebcreation/protractor-image-comparison/compare/v0.1.1...v0.1.2) (2016-11-21)
+
+
+### Bug Fixes
+
+* **_getIOSPosition:** remove app specific implementation ([62295a1](https://github.com/wswebcreation/protractor-image-comparison/commit/62295a1))
+* **baseline:** replace baselineimages for resizeDimensions ([ee82ee7](https://github.com/wswebcreation/protractor-image-comparison/commit/ee82ee7))
+* **basline:** replace baseline images for resizeDimensions ([04e1399](https://github.com/wswebcreation/protractor-image-comparison/commit/04e1399))
+* **dependencies:** set hard devdependencies ([d38fff8](https://github.com/wswebcreation/protractor-image-comparison/commit/d38fff8))
+* **nativeWebScreenshot:** fix nativeWebScreenshot only overrule if value is false ([94e0636](https://github.com/wswebcreation/protractor-image-comparison/commit/94e0636))
+* **resizeDimensions:** fix false statement ([f3d819e](https://github.com/wswebcreation/protractor-image-comparison/commit/f3d819e))
+
+
+
 <a name="0.1.1"></a>
 ## [0.1.1](https://github.com/wswebcreation/protractor-image-comparison/compare/v0.1.0...v0.1.1) (2016-11-10)